Make distance fields rendering work with Opentype CFF fonts
If the font has a CFF table, GDI will not label it as
TMPF_TRUETYPE, however, we can still use GetFontData to get
the SFNT tables. This is required to get the maxp table which
contains the glyph count, which is required to use the font
with the distance-field renderer.

Task-number: QTBUG-28746
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>
Change-Id: I1e084f7e9dbd0bccb9b1ff4de2eaf65d6a5f9f1e
(cherry picked from qtbase/dde09c429ae8b7ad0df4e4b36b8459d2b85a1219)
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>