Add newline after stream content as per PDF specification
PDF specification 7.3.8.1 says that there should be a newline after
stream data and before the endstream and this newline should not be
included in the stream length.

Before this patch the code generates a single newline (which comes
from createToUnicodeMap()) and includes it in the stream length which
causes the following the apache preflight validator error:

1.2.5 : Body Syntax error, Stream length is invalide

After this patch the code generates two newlines and includes only one
in the length.