Don't jpeg compress images that were not orginally jpeg compressed, as it generally causes non-photography images to have very visible artifacts. Add property to override this behaviour.
Change the behaviour of imageDPI such that we can set it to 0 to prevent the rescaling of images (that is, 0 is interpreted as infinity).

Don't waste time compressing the same data several times.
Don't make decisions on which kind of compression to use on the compressed image data by compressing the original data.
Fix invalid pdf files being generated in rare degenerate cases where the compressed version of the scaled image is larger than the compressed version of the unscaled data.
Don't limit the size of jpeg compressed images to 1MB.

Prepare for supporting deflated images with predictor (PNG style compression).