Calculate requiredHeight based on tallest cell
Previously the `requiredHeight` for the row was set to the height of the first cell found in the row. If another cell happened to want to push the row height taller, the row height was not correctly calculated and it appeared the row would fit on the current page. Once we get to rendering, the row _doesn't_ fit on the current page and flips itself to the next page -- without the headHeight adjustment to the row position. The THEAD then prints itself over the top of that row.

The processing time here will take longer as we now need to determine the `contentLogicalHeight` for every cell rather than just the first, but there's no way around that.